update at purchase

+++ CultureWeb/CultureWeb/Models/PurchaseDetail.cs
@@ -19,6 +19,6 @@ namespace CultureWeb.Models
         public Products Product { get; set; }
 
         public int QtyPurchase { get; set; }
-        public int CostPrice { get; set; }
+        public decimal CostPrice { get; set; }
     }
 }
